+++ src/event/sub-modules/dom/sub-modules/base/tests/specs/fire-spec.js
@@ -5,24 +5,6 @@
 KISSY.use("dom,event/dom/base", function (S, DOM, Event) {
     describe("fire", function () {
 
-        it("works for mouseenter/leave", function () {
-
-            var n = DOM.create("<div/>"), ret;
-            Event.on(n, "mouseenter", function (e) {
-                expect(e.type).toBe('mouseenter');
-                ret = 1
-            });
-            Event.fire(n, "mouseenter", {
-                relatedTarget: document
-            });
-
-            expect(ret).toBe(1);
-
-            Event.detach(n);
-
-        });
-
-
         it('support once', function () {
 
             var n = DOM.create("<div/>"), ret;